pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "vsdx-to-html"
version = "0.1.0"
description = "Konverter Visio .vsdx-filer til standalone HTML"
requires-python = ">=3.8"
dependencies = [
    "python-vsdx>=0.6.6",
]

[project.scripts]
convert = "convert:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["parser*", "renderer*"]
